+++ R/PLNnetworkfamily-class.R
@@ -50,18 +50,19 @@ PLNnetworkfamily <-
 )
 
 PLNnetworkfamily$set("public", "initialize",
-  function(penalties, responses, covariates, offsets, weights, model, control) {
+  function(penalties, responses, covariates, offsets, weights, model, xlevels, control) {
 
     ## initialize fields shared by the super class
     super$initialize(responses, covariates, offsets, weights, control)
     ## A basic model for inception
-    myPLN <- PLNfit$new(responses, covariates, offsets, weights, model, control)
+    myPLN <- PLNfit$new(responses, covariates, offsets, weights, model, xlevels, control)
     myPLN$optimize(responses, covariates, offsets, weights, control)
     control$inception <- myPLN
     ## Get an appropriate grid of penalties
     if (is.null(penalties)) {
       if (control$trace > 1) cat("\n Recovering an appropriate grid of penalties.")
-      max_pen <- max(abs(myPLN$model_par$Sigma / control$penalty_weights))
+      S <- myPLN$model_par$Sigma / control$penalty_weights
+      max_pen <- max(abs(S[upper.tri(S, diag = control$penalize_diagonal)]))
       penalties <- 10^seq(log10(max_pen), log10(max_pen*control$min.ratio), len = control$nPenalties)
     } else {
       if (control$trace > 1) cat("\nPenalties already set by the user")
@@ -71,7 +72,7 @@ PLNnetworkfamily$set("public", "initialize",
     ## instantiate as many models as penalties
     private$params <- sort(penalties, decreasing = TRUE)
     self$models <- lapply(private$params, function(penalty) {
-      PLNnetworkfit$new(penalty, responses, covariates, offsets, weights, model, control)
+      PLNnetworkfit$new(penalty, responses, covariates, offsets, weights, model, xlevels, control)
     })
 
 })
@@ -137,9 +138,10 @@ PLNnetworkfamily$set("public", "stability_selection",
                                     covariates = self$covariates[subsample, , drop = FALSE],
                                     offsets    = self$offsets   [subsample, , drop = FALSE],
                                     model      = private$model,
+                                    xlevels    = private$xlevels,
                                     weights    = self$weights   [subsample], control = ctrl_init)
 
-      ctrl_main <- PLNnetwork_param(control, inception_$n, inception_$p, inception_$d, !all(self$weights == 1))
+      ctrl_main <- PLNnetwork_param(control, inception_$n, inception_$p, inception_$d)
       ctrl_main$trace <- 0
       myPLN$optimize(ctrl_main)
       nets <- do.call(cbind, lapply(myPLN$models, function(model) {