+++ src/data_struct.h
@@ -17,9 +17,9 @@ typedef struct optim_data {
     arma::vec w          ;
     arma::mat Omega      ;
     arma::mat Theta      ;
+    double w_bar         ;
     double log_det_Omega ;
-    double KY            ;
-    arma::vec KYi        ;
+    arma::vec Ki         ;
     int iterations       ;
     int n                ;
     int p                ;
@@ -41,56 +41,10 @@ typedef struct optim_data {
         n = Y.n_rows ;
         p = Y.n_cols ;
         d = X.n_cols ;
-        iterations = 0 ;
-        KYi = logfact(Y) ;
-        KY = accu(w % KYi) ;
+        iterations = 0     ;
+        w_bar = accu(w)    ;
+        Ki = - logfact(Y) + .5 * (1+(1-p)* std::log(2*M_PI)) ;
       } ;
-    // Rank-Constrained constructor
-    optim_data(const arma::mat &responses,
-               const arma::mat &covariates,
-               const arma::mat &offsets,
-               const arma::mat &weights,
-               const int rank
-    ) : Y(responses), X(covariates), O(offsets), w(weights), q(rank)
-      {
-        n = Y.n_rows ;
-        p = Y.n_cols ;
-        d = X.n_cols ;
-        iterations = 0 ;
-        KYi = logfact(Y) ;
-        KY = accu(KYi) ;
-      } ;
-    // Sparse covariance constructor
-    optim_data(const arma::mat &responses,
-               const arma::mat &covariates,
-               const arma::mat &offsets,
-               const arma::mat &weights,
-               const arma::mat covinv
-    ) : Y(responses), X(covariates), O(offsets), w(weights), Omega(covinv), log_det_Omega(real(log_det(covinv)))
-      {
-        n = Y.n_rows ;
-        p = Y.n_cols ;
-        d = X.n_cols ;
-        iterations = 0 ;
-        KYi = logfact(Y) ;
-        KY = accu(KYi) ;
-      } ;
-    // PLN VE-step constructor
-    optim_data(const arma::mat &responses,
-               const arma::mat &covariates,
-               const arma::mat &offsets,
-               const arma::mat &regression_parameters,
-               const arma::mat &covinv,
-               const double log_det
-    ) : Y(responses), X(covariates), O(offsets), Omega(covinv), Theta(regression_parameters), log_det_Omega(log_det)
-    {
-      n = Y.n_rows ;
-      p = Y.n_cols ;
-      d = X.n_cols ;
-      iterations = 0 ;
-      KYi = logfact(Y) ;
-      KY = accu(KYi) ;
-    } ;
 
 } optim_data ;