+++ src/hessians.h
@@ -1,0 +1,8 @@
+#ifndef _hessian_H
+#define _hessian_H
+
+#include "data_struct.h"
+
+void fn_precond_PLN(unsigned N, const double *x, const double *v, double *Hv, void *data) ;
+
+#endif