+++ tests/testthat/test-plnnetworkfamily.R
@@ -13,15 +13,18 @@ test_that("PLNnetwork: main function, fields access and methods", {
   Y <- as.matrix(trichoptera$Abundance)
   O <- matrix(0, nrow(Y),ncol(Y))
   w <- rep(1, nrow(Y))
+  xlevels <- NULL
 
   ## extract the data matrices and weights
-  ctrl_main <- PLNmodels:::PLNnetwork_param(list(), nrow(Y), ncol(Y),ncol(X), weighted = FALSE)
-  ctrl_init <- PLNmodels:::PLN_param(list(), nrow(Y), ncol(Y), ncol(X), weighted = FALSE)
+  ctrl_main <- PLNmodels:::PLNnetwork_param(list(), nrow(Y), ncol(Y),ncol(X))
+  ctrl_init <- PLNmodels:::PLN_param(list(), nrow(Y), ncol(Y), ncol(X))
   ctrl_init$trace <- 0; ctrl_init$nPenalties <- 30; ctrl_init$min.ratio   <- .1
   ctrl_init$penalty_weights <- ctrl_main$penalty_weights
+  ctrl_init$penalize_diagonal <- ctrl_main$penalize_diagonal
 
   ## instantiate
-  myPLN <- PLNmodels:::PLNnetworkfamily$new(NULL, Y, X, O, w, Abundance ~ 1, ctrl_init)
+  myPLN <- PLNmodels:::PLNnetworkfamily$new(NULL, Y, X, O, w, Abundance ~ 1,
+                                            xlevels, ctrl_init)
 
   ## optimize
   myPLN$optimize(ctrl_main)